test(ts): share benchmark entrypoints for browser and native
