--- crypto-ffi/bindings/js/packages/browser/benches/addUser.bench.ts
@@ -1,9 +1,0 @@
-import { describe } from "mocha";
-import { runBenchmark } from "./utils";
-import { setupAddUserBench } from "../../../shared/benches/addUser";
-
-describe("benchmark", () => {
-    it(`Add User`, async () => {
-        await runBenchmark(setupAddUserBench);
-    });
-});
