--- crypto-ffi/bindings/js/packages/browser/benches/createMessage.bench.ts
@@ -1,9 +1,0 @@
-import { describe } from "mocha";
-import { runBenchmark } from "./utils";
-import { setupCreateMessageBench } from "../../../shared/benches/createMessage";
-
-describe("benchmark", () => {
-    it(`Create Message`, async () => {
-        await runBenchmark(setupCreateMessageBench);
-    });
-});
