--- crypto-ffi/bindings/js/packages/browser/benches/joinGroup.bench.ts
@@ -1,9 +1,0 @@
-import { describe } from "mocha";
-import { runBenchmark } from "./utils";
-import { setupJoinGroupBench } from "../../../shared/benches/joinGroup";
-
-describe("benchmark", () => {
-    it(`Join Group`, async () => {
-        await runBenchmark(setupJoinGroupBench);
-    });
-});
