--- crypto-ffi/bindings/js/packages/browser/benches/processMessage.bench.ts
@@ -1,9 +1,0 @@
-import { describe } from "mocha";
-import { runBenchmark } from "./utils";
-import { setupProcessMessageBench } from "../../../shared/benches/processMessage";
-
-describe("benchmark", () => {
-    it(`Process Message`, async () => {
-        await runBenchmark(setupProcessMessageBench);
-    });
-});
