--- crypto-ffi/bindings/js/packages/browser/benches/removeUser.bench.ts
@@ -1,9 +1,0 @@
-import { describe } from "mocha";
-import { runBenchmark } from "./utils";
-import { setupRemoveUserBench } from "../../../shared/benches/removeUser";
-
-describe("benchmark", () => {
-    it(`Remove User`, async () => {
-        await runBenchmark(setupRemoveUserBench);
-    });
-});
