--- crypto-ffi/bindings/js/packages/native/benches/addUser.bench.ts
@@ -1,8 +1,0 @@
-import { setupAddUserBench } from "../../../shared/benches/addUser";
-import { runBenchmark } from "./utils";
-
-async function run() {
-    await runBenchmark(setupAddUserBench);
-}
-
-await run();
