--- crypto-ffi/bindings/js/packages/native/benches/createMessage.bench.ts
@@ -1,8 +1,0 @@
-import { runBenchmark } from "./utils";
-import { setupCreateMessageBench } from "../../../shared/benches/createMessage";
-
-async function run() {
-    await runBenchmark(setupCreateMessageBench);
-}
-
-await run();
