--- crypto-ffi/bindings/js/packages/native/benches/joinGroup.bench.ts
@@ -1,8 +1,0 @@
-import { setupJoinGroupBench } from "../../../shared/benches/joinGroup";
-import { runBenchmark } from "./utils";
-
-async function run() {
-    await runBenchmark(setupJoinGroupBench);
-}
-
-await run();
