--- crypto-ffi/bindings/js/packages/native/benches/processMessage.bench.ts
@@ -1,8 +1,0 @@
-import { runBenchmark } from "./utils";
-import { setupProcessMessageBench } from "../../../shared/benches/processMessage";
-
-async function run() {
-    await runBenchmark(setupProcessMessageBench);
-}
-
-await run();
