--- crypto-ffi/bindings/js/packages/native/benches/removeUser.bench.ts
@@ -1,8 +1,0 @@
-import { runBenchmark } from "./utils";
-import { setupRemoveUserBench } from "../../../shared/benches/removeUser";
-
-async function run() {
-    await runBenchmark(setupRemoveUserBench);
-}
-
-await run();
